Rename basecommand -> base_command
