+++ src/pip/_internal/cli/base_command.py
