+++ src/pip/_internal/commands/check.py
@@ -1,6 +1,6 @@
 import logging
 
-from pip._internal.cli.basecommand import Command
+from pip._internal.cli.base_command import Command
 from pip._internal.operations.check import (
     check_package_set, create_package_set_from_installed,
 )
