+++ src/pip/_internal/commands/freeze.py
@@ -4,7 +4,7 @@ import sys
 
 from pip._internal import index
 from pip._internal.cache import WheelCache
-from pip._internal.cli.basecommand import Command
+from pip._internal.cli.base_command import Command
 from pip._internal.compat import stdlib_pkgs
 from pip._internal.operations.freeze import freeze
