+++ src/pip/_internal/commands/hash.py
@@ -4,7 +4,7 @@ import hashlib
 import logging
 import sys
 
-from pip._internal.cli.basecommand import Command
+from pip._internal.cli.base_command import Command
 from pip._internal.status_codes import ERROR
 from pip._internal.utils.hashes import FAVORITE_HASH, STRONG_HASHES
 from pip._internal.utils.misc import read_chunks
