+++ src/pip/_internal/vcs/__init__.py
@@ -17,7 +17,7 @@ from pip._internal.utils.typing import MYPY_CHECK_RUNNING
 
 if MYPY_CHECK_RUNNING:
     from typing import Dict, Optional, Tuple  # noqa: F401
-    from pip._internal.cli.basecommand import Command  # noqa: F401
+    from pip._internal.cli.base_command import Command  # noqa: F401
 
 __all__ = ['vcs', 'get_src_requirement']
